Support install/cleanup of Local Storage Operator
This adds configurable support for the Local Storage Operator
within our dev-scripts. It automatically attaches disks to
the configured worker VMs and configures the local-storage-operator
to provide those as PVs for the 'local' storage class.

Two Makefile targets are included:

 -make local_storage
 -make local_storage_cleanup

These targets are both currently optional (opt in) for those who
wish to experiment with an alternate PV type.