Add datavolume playbook
This playbook pre-creates an openstack-base-img DataVolume
using virtctl on the CLI. This has the effect of uploading
the image into the cluster from disk thus avoiding
repeatedly downloading the same image. This also works
around a case where cdi-datavolume streamed http downloads
can be really slow from some mirrors (more to be said about
that but perhaps not here).