Align OLM cleanup for webUI and CLI
The naming schemes for the subscription and operator group
were slightly different. This patch aligns things so that
olm_cleanup will remove the installed OLM resources regardless
of whether the CLI or UI is used to install them.

The operatorgroup resource gets auto-created with a generated
name via the UI so we add logic to the playbook to lookup and use
that for deletion accordingly.