Fix cidr_suffix extract + ipv6 multi subnet
The cidr_suffix was not correct extracted from the subnet cidr
to create the gatway interface on the host. This fixes it and
also adds a task to create IPv6 gateway interface if the subnet
is IPv6.

Also moves the role_data.yaml to be a template that we only have
one file per version to maintain. Before there were two as we
had a separate one for IPv6 to add IPv6 and IPv4 gateway for the
controller role.