[OSPK8-449] Automate compose/puddle information based on input file
Adds osp_release_auto_url var which can be used to point to the
compose container_image_prepare.yaml. From there information is
parsed on the RHEL relase and containers to use to deploy OSP.

With this it would be no longer required to manually bump container
tags when a new compos is available.

Per default osp_release_auto_url points to the OSP16.2 passed_phase2
compose url.