[OSPK8-451] Move ctlplane network to a default MTU
The ospnetwork has a MTU of 9000 that it is possible to also use/test
jumbo frames with the storage/tenant network. The downstream repo
service does not/stopped to respond to ICMP, therefore pmtu does not
work properly and installing packages after ooo network config got
applied fail on the computes nodes, where the external network with
its default gw is not configured.

This moves the ctlplane network to the metal3 provisioning nic
enp1s0 which has default 1500 MTU.