Fix osp_release_auto
The osp_release_auto yaml can persist been runs.
'make openstack_cleanup' will remove it, however 'make olm_cleanup' will
create it again as it includes the local role.
On the next deployment the existing release vars will be used instead of
the correct values for OSP_RELEASE.

For now just force it to be downloaded every time to ensure it does the
right thing.